tests/wpt/web-platform-tests/html/cross-origin-opener-policy/reporting-coop-navigated-popup.https.html

<title>Cross-Origin-Opener-Policy: a navigated popup with reporting</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/utils.js"></script> <!-- Use token() to allow running tests in parallel -->
<script src="resources/reporting-common.js"></script>

<script>

// This test does the following:
//  1 - This document has COOP: same-origin-allow-popups; report-to="coop-report-endpoint"
//  2 - Open a popup on a same-origin page without COOP, with the coop-popup-report-endpoint
//  3 - Navigate the popup to a same-origin page with COOP, with the coop-redirect-report-endpoint
// it verifies that the reports are properly send for the browsing context switch
// during the navigation in the popup (step 3). The current document (the opener)
// endpoint should not receive any report as no switch ocurred on 2.
promise_test( async t => {
  const popupName = token();
  const noCoopChannelName = token();
  const coopChannelName = token();
  await reportingTest( (resolve) => {
    const noCOOP = `resources/coop-coep.py?coop=${encodeURIComponent(`unsafe-none; report-to="${popupReportEndpoint.name}"`)}&coep=&channel=${noCoopChannelName}`;
    const coop = `resources/coop-coep.py?coop=${encodeURIComponent(`same-origin; report-to="${redirectReportEndpoint.name}"`)}&coep=&channel=${coopChannelName}`;

    const popup = window.open(noCOOP, popupName);
    const channel = new BroadcastChannel(coopChannelName);
    // Close the popup once the test is complete.
    // The browsing context is closed after the navigation hence use the
    // broadcast channel to trigger the closure.
    t.add_cleanup(() => {
      channel.postMessage("close");
    });
    popup.onload = t.step_func(() => {
      assert_equals(popup.name.length, popupName.length, "popup name");
      channel.onmessage = t.step_func(event => {
        const payload = event.data;
        // The name should be empty, but we're checking the length rather than a
        // string comparison to "" to keep the random token out of error messages.
        assert_equals(payload.name.length, 0, "Popup name after navigation");
        assert_false(payload.opener, "Opener after navigation");
        assert_true(popup.closed, "Window proxy closed after navigation");
        resolve();
      });
      popup.location = coop;
    });
  },
  popupName,
  [
    // Reports expected for the navigation from "noCOOP" to "coop"
    {
      "endpoint": popupReportEndpoint,
      "report": {
        "body": {
          "disposition": "enforce",
          "document-uri": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
          "effective-policy": "unsafe-none",
          "navigation-uri": RegExp(`coop-coep.py?.*channel=${coopChannelName}$`),
          "violation-type": "navigation-from-document"
        },
        "url": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
        "type": "coop"
      }
    },
    {
      "endpoint": redirectReportEndpoint,
      "report": {
        "body": {
          "disposition": "enforce",
          "document-uri": RegExp(`coop-coep.py?.*channel=${coopChannelName}$`),
          "effective-policy": "same-origin",
          "navigation-uri": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
          "violation-type": "navigation-to-document"
        },
        "url": RegExp(`coop-coep.py?.*channel=${coopChannelName}$`),
        "type": "coop"
      }
    },
  ]);
}, "Open a popup to a document without COOP, then navigate it to a document with");

verifyRemainingReports();

</script>
